Update README.md
TESLA shares have been quite volatile since the company’s foundation in 2010. If you had all the companies share data for the last 10 years, can you make a prediction regarding its future value? Requirements You are required to take the dataset “TSLA.csv” and using any Machine Leaning Algorithm, make a prediction as to what the stock’s price will be on the 03/03/2020. (The data set contains 10 years’ worth of data finishing on the 03/02/2020) The output of your program must:

Visualize the data and prediction
Print out the message: “The predicted share price on the 3rd March 2020 is” (predicted share price)
Print out the accuracy or error of the ML model.
Explain in the code (As a comment) why you picked the specific algorithm